+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 
 BREAKING CHANGES
 
+- [core] Set validators limit to 100 for testnet
 - [tendermint] Update to [v0.26.0](https://github.com/tendermint/tendermint/blob/master/CHANGELOG.md#v0260) 
+
+IMPROVEMENT
+
+- [config] Add keep_state_history option
 
 ## 0.5.1
 *Oct 22th, 2018*